+++ changelog.md
@@ -6,6 +6,7 @@
 - New: basic.print (improves and supersedes basic.float, basic.formula and basic.value)
 - New: basic.color (supersedes basic.colordisc and basic.rgb)
 - New: device.uzsuicon (to control UZSU in smarthome.py and FHEM)
+- New: calendar.waste (compact view of waste collection dates)
 - Enhanced plot.period, among other things merged functionality of plot.minmaxavg and plot.multiaxis into it
 - plot.temprose: New parameters 'series_label' and 'unit'
 - basic.slider: New parameter 'value_display'